refator : 컴포넌트 분리 및 사진을 Link태그로 감싸서 상세페이지와 연결(#24)
