+++ src/types/funding.ts
@@ -1,0 +1,8 @@
+export type campaignProps = {
+    id: string;
+    targetAmount: number;
+    currentAmount: number;
+    deadline: Date;
+    photo: string;
+    title: string;
+  };
